+++ app/models/formatters/imperial_height_formatter.rb
@@ -5,7 +5,7 @@ module Formatters
   class ImperialHeightFormatter < IntegerFormatter
     include DateAndTimeParser
 
-    def name_response(response)
+    def name_response(response, shared_responses = domain_options)
       hash = parse_imperial_height(response)
       f = (hash[:feet] == 1 ? 'foot' : 'feet')
       i = (hash[:inches] == 1 ? 'inch' : 'inches')