+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.52.2
 
+### Bug Fix
+- Fixed a bug that prevented labeled CSV exports from finishing on projects that
+  collected imperial height and imperial weight
+
 ## 0.52.1 (May 12, 2017)
 
 ### Bug Fix